Remove section on extending.
I don't like to promote syntax extensions in the first place because they seem
to generally make life harder.  However, I also don't have an example use case
for this, and I have reason to doubt this would work.  The syntax extensions
wouldn't run natively in an uncompiled ".mjs" file.  The tool would need to
output both an ".mjs" and a ".js" file for that to work, which would be more
complex, and I'd rather avoid complexity.